Move metrics handler utilities to new package in container-core
