+++ container-core/src/main/java/com/yahoo/container/handler/metrics/package-info.java
@@ -1,0 +1,6 @@
+// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+/** Exported config package */
+@ExportPackage
+package com.yahoo.container.handler.metrics;
+
+import com.yahoo.osgi.annotation.ExportPackage;
