+++ container-core/src/test/java/com/yahoo/container/handler/metrics/ErrorResponseTest.java
@@ -1,5 +1,5 @@
 // Copyright 2020 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.metricsproxy.http;
+package com.yahoo.container.handler.metrics;
 
 import org.junit.Test;
